requirements.txt
requests
opencv-python
# gstreamer-python
onvif-zeep
asyncio
aiohttp
pydantic
loguru
